Handle comment differently
